refactor: :recycle: Inline downloaders to the download manager, it's more convenient to work with progress that way
